consolidate icon type into existing icon type
