--- src/app/shared/models/buttons/icon-type.ts
@@ -1,1 +1,0 @@
-export type IconType = 'export' | 'dots' | undefined;
