+++ src/app/shared/models/icon-type.ts
@@ -1,0 +1,1 @@
+export type IconType = 'export' | 'dots' |  'document' | 'organization' | 'systems' | 'clipboard' | 'folder-important' | 'bulk-create' | undefined;
